+++ ra_cpp/include/isleft.h
@@ -11,8 +11,8 @@
 
 namespace py = pybind11;
 
-double isleft(Eigen::Ref<Eigen::RowVector2d> Vx, 
-    Eigen::Ref<Eigen::RowVector2d> Vy, 
-    Eigen::Ref<Eigen::RowVector2d> ref_point2d);
+double isleft(Eigen::Ref<Eigen::RowVector2f> Vx,
+    Eigen::Ref<Eigen::RowVector2f> Vy,
+    Eigen::Ref<Eigen::RowVector2f> ref_point2d);
 
 #endif /* ISLEFT_H */
