+++ ra_cpp/include/refpoint.h
@@ -11,9 +11,9 @@
 
 namespace py = pybind11;
 
-Eigen::RowVector3d refpoint(Eigen::Ref<Eigen::RowVector3d> ray_origin, 
-        Eigen::Ref<Eigen::RowVector3d> v_in,
-        Eigen::Ref<Eigen::RowVector3d> normal,
-        Eigen::Ref<Eigen::RowVector3d> vertcoord);
+Eigen::RowVector3f refpoint(Eigen::Ref<Eigen::RowVector3f> ray_origin,
+        Eigen::Ref<Eigen::RowVector3f> v_in,
+        Eigen::Ref<Eigen::RowVector3f> normal,
+        Eigen::Ref<Eigen::RowVector3f> vertcoord);
 
 #endif /* REFPOINT_H */
